# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Wavelet 平台后端 API，提供用户认证、系统配置、任务调度等通用功能。",
        "title": "Wavelet API",
        "contact": {
            "name": "Wavelet",
            "url": "https://github.com/Rain-kl/Wavelet"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/admin/system-configs": {
            "get": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "返回所有系统配置列表，支持按配置类型（system/business）过滤，需要管理员权限",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "获取系统配置列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "配置类型（system/business）",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "系统配置列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseAny"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.SystemConfig"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    },
                    "403": {
                        "description": "无管理员权限",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "创建一条新的系统配置项，配置键不可重复，同时将新配置同步到 Redis，需要管理员权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "创建系统配置",
                "parameters": [
                    {
                        "description": "创建请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/system_config.CreateSystemConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseAny"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误或配置键已存在",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    },
                    "403": {
                        "description": "无管理员权限",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/system-configs/smtp/test": {
            "post": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "使用传入的配置进行 SMTP 邮件发送测试，支持使用 ****** 占位符使用保存的数据库密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "测试 SMTP 邮件发送",
                "parameters": [
                    {
                        "description": "测试请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/system_config.TestSMTPRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "测试执行完毕",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseAny"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/system_config.TestSMTPResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/system-configs/{key}": {
            "get": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "根据配置键获取对应的系统配置详情，需要管理员权限",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "获取单个系统配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "配置键",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "系统配置详情",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseAny"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.SystemConfig"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    },
                    "403": {
                        "description": "无管理员权限",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    },
                    "404": {
                        "description": "配置不存在",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "根据配置键更新对应的配置内容，同时将更新同步到 Redis，需要管理员权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "更新系统配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "配置键",
                        "name": "key",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/system_config.UpdateSystemConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseAny"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    },
                    "403": {
                        "description": "无管理员权限",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    },
                    "404": {
                        "description": "配置不存在",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.SystemConfig": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                },
                "visibility": {
                    "type": "integer"
                }
            }
        },
...
        "system_config.CreateSystemConfigRequest": {
            "type": "object",
            "required": [
                "key",
                "type",
                "value"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 255
                },
                "key": {
                    "type": "string",
                    "maxLength": 64
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "system",
                        "business"
                    ]
                },
                "value": {
                    "type": "string",
                    "maxLength": 255
                },
                "visibility": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                }
            }
        },
        "system_config.TestSMTPRequest": {
            "type": "object",
            "required": [
                "smtp_host",
                "smtp_password",
                "smtp_port",
                "smtp_username",
                "to"
            ],
            "properties": {
                "smtp_host": {
                    "type": "string",
                    "maxLength": 255
                },
                "smtp_password": {
                    "type": "string",
                    "maxLength": 255
                },
                "smtp_port": {
                    "type": "integer"
                },
                "smtp_username": {
                    "type": "string",
                    "maxLength": 255
                },
                "to": {
                    "type": "string"
                }
            }
        },
        "system_config.TestSMTPResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "log": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "system_config.UpdateSystemConfigRequest": {
            "type": "object",
            "required": [
                "value"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 255
                },
                "value": {
                    "type": "string",
                    "maxLength": 255
                },
                "visibility": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                }
            }
        },
...
        "util.ResponseAny": {
            "type": "object",
            "properties": {
                "data": {},
                "error_msg": {
                    "type": "string",
                    "example": ""
                }
            }
        }
    },
    "securityDefinitions": {
        "SessionCookie": {
            "type": "apiKey",
            "name": "session",
            "in": "cookie"
        }
    }
}